backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
boto3==1.34.14
azure-cognitiveservices-speech==1.48.2
python-multipart==0.0.6
httpx==0.25.1
pydantic==2.12.5
pydub==0.25.1
librosa==0.10.0
sentence-transformers==2.7.0
sarvamai==0.1.27
